requirements_dev.txt
pyserial
click
pydantic
pyyaml
pyside6


# Development
pytest
mypy
black
vulture
ruff
rich
pylint
ruff

# Creating documentation
sphinx
sphinx_click
furo

# Creating an installer
pyinstaller
nuitka